+++ server/component/cms/CmsView.php
@@ -612,6 +612,17 @@ class CmsView extends BaseView
                 "items" => $this->model->get_db()->fetch_table_as_select_values('qualtricsSurveys', 'id', array('name', 'qualtrics_survey_id'))
             ));
         }
+        else if($field['type'] == "select-formName")
+        {
+            $children[] = new BaseStyleComponent("select", array(
+                "value" => $field['content'],
+                "name" => $field_name_prefix . "[content]",
+                "max" => 10,
+                "live_search" => 1,
+                "is_required" => 1, 
+                "items" => $this->model->get_db()->fetch_table_as_select_values('view_form', 'form_id', array('form_name'))
+            ));
+        }
         else if($field['type'] == "select-plugin")
         {
             $children[] = new BaseStyleComponent("select", array(
@@ -691,6 +702,14 @@ class CmsView extends BaseView
                 "name" => $field['name'],
                 "disabled" => 1,
                 "items" => $this->model->get_db()->fetch_table_as_select_values('qualtricsSurveys', 'id', array('name', 'qualtrics_survey_id'))
+            ));
+        }
+        else if ($field['type'] == "select-formName") {
+            $children[] = new BaseStyleComponent("select", array(
+                "value" => $field['content'],
+                "name" => $field['name'],
+                "disabled" => 1,
+                "items" => $this->model->get_db()->fetch_table_as_select_values('view_form', 'form_id', array('form_name'))
             ));
         }
         else if($field['type'] == "select-plugin")