+++ server/component/style/StyleView.php
@@ -95,6 +95,25 @@ abstract class StyleView extends BaseView
     }
 
     /**
+     * Render the content of all children of this view instance as entries
+     * @param array $entry_value
+     * the data for the entry value
+     */
+    protected function output_children_entry($entry_data)
+    {
+        foreach ($this->children as $child)
+        if ($child instanceof StyleComponent || $child instanceof BaseStyleComponent) {
+            if(method_exists($child, 'output_content_entry')){
+                $child->output_content_entry($entry_data);
+            }else{
+                $child->output_content();
+            }
+        } else {
+            echo "invalid child element of type '" . gettype($child) . "'";
+        };
+    }
+
+    /**
      * Render the content of all children of this view instance.
      */
     protected function output_children_mobile()
@@ -103,6 +122,28 @@ abstract class StyleView extends BaseView
         foreach ($this->children as $child) {
             if ($child instanceof StyleComponent || $child instanceof BaseStyleComponent) {
                 $res[] = $child->output_content_mobile();
+            } else {
+                echo "invalid child element of type '" . gettype($child) . "'";
+            }
+        }
+        return $res;
+    }
+
+    /**
+     * Render the content of all children of this view instance as entries
+     * @param array $entry_value
+     * the data for the entry value
+     */
+    protected function output_children_mobile_entry($entry_data)
+    {
+        $res = [];
+        foreach ($this->children as $child) {
+            if ($child instanceof StyleComponent || $child instanceof BaseStyleComponent) {
+                if (method_exists($child, 'output_content_mobile_entry')) {
+                    $res[] = $child->output_content_mobile_entry($entry_data);
+                } else {
+                    $res[] = $child->output_content_mobile();
+                }
             } else {
                 echo "invalid child element of type '" . gettype($child) . "'";
             }
@@ -125,6 +166,33 @@ abstract class StyleView extends BaseView
             $style['fail_msgs']  = $fail_msgs;
         }
         return $style;
+    }
+
+    public function output_content_mobile_entry($entry_data)
+    {
+        $style = $this->model->get_db_fields();
+        $style['style_name'] = $this->style_name;
+        $style['css'] = $this->css;
+        $style['children'] = $this->output_children_mobile_entry($entry_data);
+        $success_msgs = $this->output_controller_alerts_success_mobile();
+        if($success_msgs){
+            $style['success_msgs'] = $success_msgs;
+        }
+        $fail_msgs = $this->output_controller_alerts_fail_mobile();
+        if($fail_msgs){
+            $style['fail_msgs']  = $fail_msgs;
+        }
+        return $style;
+    }
+
+    public function get_entry_param($input){
+        preg_match_all('~\$\w+\b~', $input, $m);
+        foreach ($m as $key => $value) {
+            if ($value) {
+                $param_name = str_replace('$', '', $value[0]);
+                return $param_name;
+            }
+        }
     }
 }
 ?>