+++ server/component/style/heading/HeadingView.php
@@ -52,6 +52,38 @@ class HeadingView extends StyleView
         if($this->level > 6) $this->level = 6;
         require __DIR__ . "/tpl_title.php";
     }
+
+    /**
+     * Render output as an entry
+     * @param array $entry_value
+     * the data for the entry value
+     */
+    public function output_content_entry($entry_value)
+    {
+        $this->title = $this->model->get_db_field("title");
+        $param = $this->get_entry_param($this->title);
+        $this->title = isset($entry_value[$param]) ? str_replace('$' . $param, $entry_value[$param], $this->title) : $this->title; // if the param is not set, return the original
+        if ($this->level < 1) {
+            $this->level = 1;
+        };
+        if ($this->level > 6) {
+            $this->level = 6;
+        };
+        require __DIR__ . "/tpl_title.php";
+    }
+
+     /**
+     * Render output as an entry for mobile
+     * @param array $entry_value
+     * the data for the entry value
+     */
+    public function output_content_mobile_entry($entry_value)
+    {
+        $style = parent::output_content_mobile();
+        $param = $this->get_entry_param($this->title);
+        $style['title']['content'] = isset($entry_value[$param]) ? str_replace('$' . $param, $entry_value[$param], $this->title) : $this->title; // if the param is not set, return the original
+        return $style;
+    }
 	
 }
 ?>