+++ server/component/style/markdown/MarkdownView.php
@@ -37,10 +37,10 @@ class MarkdownView extends StyleView
      */
     public function __construct($model)
     {
-        parent::__construct($model);        
+        parent::__construct($model);
         $this->text_md = $this->model->get_db_field('text_md');
         $this->data_config = $this->model->get_db_field("data_config");
-        if($this->data_config){
+        if ($this->data_config) {
             $this->retrieve_data();
         }
     }
@@ -50,7 +50,8 @@ class MarkdownView extends StyleView
     /**
      * Retrieve data from database - base dont the JSON configuration
      */
-    private function retrieve_data(){
+    private function retrieve_data()
+    {
         $fields = $this->model->retrieve_data($this->data_config);
         if ($fields) {
             foreach ($fields as $field_name => $field_value) {
@@ -68,25 +69,60 @@ class MarkdownView extends StyleView
      */
     public function output_content()
     {
-        if(is_a($this->model, "BaseStyleModel"))
-        {
+        if (is_a($this->model, "BaseStyleModel")) {
             $pd = new ParsedownExtension();
             $md = $pd->text($this->text_md);
-        }
-        else
+        } else
             $md = $this->text_md;
+        require __DIR__ . "/tpl_markdown.php";
+    }
+
+    /**
+     * Render output as an entry
+     * @param array $entry_value
+     * the data for the entry value
+     */
+    public function output_content_entry($entry_value)
+    {
+        $param = $this->get_entry_param($this->text_md);
+        $this->text_md = str_replace('<p>', '', $this->text_md);
+        $this->text_md = str_replace('</p>', '', $this->text_md);
+        $entry_value = isset($entry_value[$param]) ? str_replace('$' . $param, $entry_value[$param], $this->text_md) : $this->text_md; // if the param is not set, return the original
+        $pd = new ParsedownExtension();
+        $md = $pd->text($entry_value);
         require __DIR__ . "/tpl_markdown.php";
     }
 
     public function output_content_mobile()
     {
         $style = parent::output_content_mobile();
-        if($this->data_config){
+        if ($this->data_config) {
             $this->retrieve_data();
             $style['text_md']['content'] = $this->text_md;
         }
         return $style;
     }
-	
+
+    /**
+     * Render output as an entry for mobile
+     * @param array $entry_value
+     * the data for the entry value
+     */
+    public function output_content_mobile_entry($entry_value)
+    {
+        $style = parent::output_content_mobile();
+        $param = $this->get_entry_param($this->text_md);
+        $this->text_md = str_replace('<p>', '', $this->text_md);
+        $this->text_md = str_replace('</p>', '', $this->text_md);
+        $entry_value = isset($entry_value[$param]) ? str_replace('$' . $param, $entry_value[$param], $this->text_md) : $this->text_md; // if the param is not set, return the original
+        $pd = new ParsedownExtension();
+        $md = $pd->text($entry_value);
+        $style['text_md']['content'] = $md;
+        if ($this->data_config) {
+            $this->retrieve_data();
+            $style['text_md']['content'] = $md;
+        }
+        return $style;
+    }
 }
 ?>